Move tex_scripts.rs content into tex_math.rs (Perl: TeX_Math.pool.ltxml)
In Perl, sub/superscript handling (scriptHandler, is_script, revert_script,
script_sizer, and the script constructors) are all part of TeX_Math.pool.ltxml.
The Rust-only tex_scripts.rs file is now eliminated by inlining its content
into tex_math.rs, matching the Perl organization.

413 tests pass.
